+++ lib/src/clients/base/couchdb_base_client.dart
@@ -23,7 +23,7 @@ abstract class CouchDbBaseClient {
 
   /// DELETE method
   Future<DbResponse> delete(String path,
-      {Map<String, String> reqHeaders, Map<String, Object> body});
+      {Map<String, String> reqHeaders});
 
   /// COPY method
   Future<DbResponse> copy(String path, {Map<String, String> reqHeaders});