+++ lib/src/clients/couchdb_web_client.dart
@@ -1,71 +1,223 @@
-// import 'base/couchdb_base_client.dart';
+import 'dart:convert';
+import 'dart:html';
+
+import 'package:http/browser_client.dart';
+
+import '../entities/db_response.dart';
+import '../exceptions/couchdb_exception.dart';
+import 'base/couchdb_base_client.dart';
 
 /// Client for interacting with database via browser-side applications
-// class CouchDbWebClient extends CouchDbBaseClient {
-//   @override
-//   // TODO: implement connectUri
-//   String get connectUri => null;
+class CouchDbWebClient extends CouchDbBaseClient {
+  /// Creates only one instance of [CouchDbWebClient] per multiple calls.
+  /// 
+  /// It is recommend to don't use CouchDB as standalone server, because CouchDB is created for one thing -
+  /// storing data and no more. Though it is your decision.
+  factory CouchDbWebClient(
+          {String username,
+          String password,
+          String host = '127.0.0.1',
+          int port = 5984}) =>
+      _client ??= CouchDbWebClient._create(username, password, host, port);
 
-//   @override
-//   Future<Map<String, Object>> copy(String path) {
-//     // TODO: implement copy
-//     return null;
-//   }
+  CouchDbWebClient._create(
+      this.username, this.password, this.host, this.port);
 
-//   @override
-//   Future<Map<String, Object>> delete(String path, [Object body]) {
-//     // TODO: implement delete
-//     return null;
-//   }
+  static CouchDbWebClient _client;
 
-//   @override
-//   Future<Map<String, Object>> get(String path) {
-//     // TODO: implement get
-//     return null;
-//   }
+  /// Host
+  String host;
 
-//   @override
-//   Future<Map<String, List<String>>> head(String path) {
-//     // TODO: implement head
-//     return null;
-//   }
+  /// Port
+  int port;
 
-//   @override
-//   Future<Map<String, Object>> post(String path, Object body) {
-//     // TODO: implement post
-//     return null;
-//   }
+  /// Username of database user
+  String username;
 
-//   @override
-//   Future<Map<String, Object>> put(String path, {Object body}) {
-//     // TODO: implement put
-//     return null;
-//   }
+  /// Password of database user
+  String password;
 
-//   factory CouchDbWebClient({
-//     String username,
-//     String password,
-//     String host = '127.0.0.1',
-//     int port = 5984
-//   }) => _client ??= CouchDbWebClient._create(
-//           username,
-//           password,
-//           host,
-//           port
-//         );
+  final _browserClient = BrowserClient();
 
-//   CouchDbWebClient._create(
-//     this.username,
-//     this.password,
-//     this.host,
-//     this.port
-//   );
+  /// Base64 encoded [username] and [password]
+  String get authCredentials => const Base64Encoder().convert('$username:$password'.codeUnits);
 
-//   static CouchDbWebClient _client;
+  @override
+  String get connectUri => 'http://$host:$port';
 
-//   String host;
-//   int port;
-//   String username;
-//   String password;
+  @override
+  Future<DbResponse> head(String path, {Map<String, String> reqHeaders}) async {
+    final resHeaders = <String, List<String>>{};
 
-// }
+    final tmpHeaders = reqHeaders ?? <String, String>{};
+    tmpHeaders['Authorization'] = 'Basic $authCredentials';
+
+    final res = await _browserClient.head('$connectUri/$path', headers: tmpHeaders);
+    _browserClient.close();
+
+    if (res.statusCode != 200 &&
+        res.statusCode != 201 &&
+        res.statusCode != 202) {
+      throw CouchDbException(res.statusCode);
+    }
+
+    // TODO(YevenKap): split headerValue
+    res.headers.forEach((headerName, headerValue) => resHeaders[headerName] = <String>[headerValue]);
+    return DbResponse(headers: resHeaders);
+  }
+
+  @override
+  Future<DbResponse> get(String path, {Map<String, String> reqHeaders}) async {
+    final resHeaders = <String, List<String>>{};
+    Map<String, Object> json;
+
+    final tmpHeaders = reqHeaders ?? <String, String>{};
+    tmpHeaders['Authorization'] = 'Basic $authCredentials';
+    tmpHeaders.putIfAbsent('Accept', () => 'application/json');
+
+    final res = await _browserClient.get('$connectUri/$path', headers: tmpHeaders);
+    _browserClient.close();
+
+    switch (res.headers['content-type']) {
+      case 'text/plain':
+        json = <String, Object>{'limit': int.tryParse(res.body)};
+        break;
+      case 'application/json':
+        json = Map<String, Object>.from(jsonDecode(res.body));
+        break;
+      default:
+        json = <String, Object>{'rawBody': res.body};
+    }
+
+    res.headers.forEach((headerName, headerValue) => resHeaders[headerName] = <String>[headerValue]);
+    json['headers'] = resHeaders;
+
+    if (res.statusCode != 200 &&
+        res.statusCode != 201 &&
+        res.statusCode != 202) {
+      throw CouchDbException(res.statusCode, response: DbResponse.fromJson(json));
+    }
+
+    return DbResponse.fromJson(json);
+  }
+
+  @override
+  Future<DbResponse> put(String path,
+      {Object body, Map<String, String> reqHeaders}) async {
+    final resHeaders = <String, List<String>>{};
+
+    final tmpHeaders = reqHeaders ?? <String, String>{};
+    tmpHeaders['Authorization'] = 'Basic $authCredentials';
+    tmpHeaders.putIfAbsent('Accept', () => 'application/json');
+
+    final encodedData = body is Map ? jsonEncode(body) : body;
+
+    final res = await _browserClient.put('$connectUri/$path', headers: tmpHeaders, body: encodedData);
+    _browserClient.close();
+
+    final resBody = jsonDecode(res.body);
+    final json = Map<String, Object>.from(resBody);
+
+    res.headers.forEach((headerName, headerValue) => resHeaders[headerName] = <String>[headerValue]);
+    json['headers'] = resHeaders;
+
+    if (res.statusCode != 200 &&
+        res.statusCode != 201 &&
+        res.statusCode != 202) {
+      throw CouchDbException(res.statusCode, response: DbResponse.fromJson(json));
+    }
+
+    return DbResponse.fromJson(json);
+  }
+
+  @override
+  Future<DbResponse> post(String path,
+      {Map<String, Object> body, Map<String, String> reqHeaders}) async {
+    final resHeaders = <String, List<String>>{};
+
+    final tmpHeaders = reqHeaders ?? <String, String>{};
+    tmpHeaders['Authorization'] = 'Basic $authCredentials';
+    tmpHeaders.putIfAbsent('Accept', () => 'application/json');
+
+    final encodedData = jsonEncode(body);
+
+    final res = await _browserClient.post('$connectUri/$path', headers: tmpHeaders, body: encodedData);
+    _browserClient.close();
+
+    final resBody = jsonDecode(res.body);
+
+    Map<String, Object> json;
+    switch (resBody.runtimeType) {
+      case List:
+        json = <String, List<Object>>{'result': List<Object>.from(resBody)};
+        break;
+      default:
+        json = Map<String, Object>.from(resBody);
+    }
+
+    res.headers.forEach((headerName, headerValue) => resHeaders[headerName] = <String>[headerValue]);
+    json['headers'] = resHeaders;
+
+    if (res.statusCode != 200 &&
+        res.statusCode != 201 &&
+        res.statusCode != 202) {
+      throw CouchDbException(res.statusCode, response: DbResponse.fromJson(json));
+    }
+
+    return DbResponse.fromJson(json);
+  }
+
+  @override
+  Future<DbResponse> delete(String path,
+      {Map<String, String> reqHeaders}) async {
+    final resHeaders = <String, List<String>>{};
+
+    final tmpHeaders = reqHeaders ?? <String, String>{};
+    tmpHeaders['Authorization'] = 'Basic $authCredentials';
+    tmpHeaders.putIfAbsent('Accept', () => 'application/json');
+
+    final res = await _browserClient.delete('$connectUri/$path', headers: tmpHeaders);
+    _browserClient.close();
+
+    final resBody = jsonDecode(res.body);
+    final json = Map<String, Object>.from(resBody);
+
+    res.headers.forEach((headerName, headerValue) => resHeaders[headerName] = <String>[headerValue]);
+    json['headers'] = resHeaders;
+
+    if (res.statusCode != 200 &&
+        res.statusCode != 201 &&
+        res.statusCode != 202) {
+      throw CouchDbException(res.statusCode, response: DbResponse.fromJson(json));
+    }
+
+    return DbResponse.fromJson(json);
+  }
+
+  @override
+  Future<DbResponse> copy(String path, {Map<String, String> reqHeaders}) async {
+    final resHeaders = <String, List<String>>{};
+
+    final tmpHeaders = reqHeaders ?? <String, String>{};
+    tmpHeaders['Authorization'] = 'Basic $authCredentials';
+    tmpHeaders.putIfAbsent('Accept', () => 'application/json');
+
+    final res = await HttpRequest.request('$connectUri/$path', requestHeaders: tmpHeaders,
+      method: 'COPY');
+
+    final rawBody = res.responseText;
+    final resBody = jsonDecode(rawBody);
+    final json = Map<String, Object>.from(resBody);
+
+    res.responseHeaders.forEach((headerName, headerValue) => resHeaders[headerName] = <String>[headerValue]);
+    json['headers'] = resHeaders;
+
+    if (res.status != 200 &&
+        res.status != 201 &&
+        res.status != 202) {
+      throw CouchDbException(res.status, response: DbResponse.fromJson(json));
+    }
+
+    return DbResponse.fromJson(json);
+  }
+}